chore: deprecate pay_to_email
As we are moving towards users possibly having access to multiple
merchant accounts, `pay_to_email` no longer makes sense.
This is a breaking change that removes it as an option from the SDK.
We already removed it from docs, APIs, and SDKs for other languages,
PHP was lacking behind.

Also fixes other parts of the documentation.

Fixes: https://github.com/sumup/sumup-ecom-php-sdk/issues/30
Fixes: https://github.com/sumup/sumup-ecom-php-sdk/issues/46
Fixes: https://github.com/sumup/sumup-ecom-php-sdk/issues/28