feat: codegen
Introduce codegen tool like we have in other SDKs to ensure we
can keep the SDK up-to-date with latest changes in our APIs
and cover all of the APIs availabel to developers.

The setup is just the same as for sumup-py or sumup-go.
For now, we are keeping the implementation as similar to the
previous version. Once there are breaking changes we might bump
the SDK to v2.